Move scm functions to a dedicated file
