+++ pkg/core/engine/scm.go
@@ -1,0 +1,72 @@
+package engine
+
+import (
+	"sync"
+
+	"github.com/mitchellh/hashstructure"
+	"github.com/sirupsen/logrus"
+	"github.com/updatecli/updatecli/pkg/core/pipeline/scm"
+)
+
+// InitSCM search and clone only once SCM configurations found.
+func (e *Engine) InitSCM() (err error) {
+	hashes := []uint64{}
+
+	wg := sync.WaitGroup{}
+	channel := make(chan int, 20)
+	defer wg.Wait()
+
+	for _, pipeline := range e.Pipelines {
+		for _, s := range pipeline.SCMs {
+
+			if s.Handler != nil {
+				err = Clone(&s.Handler, channel, &hashes, &wg)
+				if err != nil {
+					return err
+				}
+			}
+		}
+	}
+	logrus.Infof("\nSCM repository retrieved: %d", len(hashes))
+
+	return err
+}
+
+// Clone parses a scm configuration then clone the git repository if needed.
+func Clone(
+	s *scm.ScmHandler,
+	channel chan int,
+	hashes *[]uint64,
+	wg *sync.WaitGroup) error {
+
+	scmhandler := *s
+
+	hash, err := hashstructure.Hash(scmhandler.GetDirectory(), nil)
+	if err != nil {
+		return err
+	}
+	found := false
+
+	for _, h := range *hashes {
+		if h == hash {
+			found = true
+		}
+	}
+
+	if !found {
+		*hashes = append(*hashes, hash)
+		wg.Add(1)
+		go func(s scm.ScmHandler) {
+			channel <- 1
+			defer wg.Done()
+			_, err := s.Clone()
+			if err != nil {
+				logrus.Errorf("err - %s", err)
+			}
+		}(scmhandler)
+		<-channel
+
+	}
+
+	return nil
+}
